feat: refactor test fixtures
